requirements.txt
numpy
langchain
langchain-community
langchain-core
langchain-openai
chromadb
PyPDF2
sentence-transformers
faiss-cpu
gradio
openai
#sqlite3
psycopg2-binary
fastapi
uvicorn
httpx
pydantic
python-dotenv
